refactor: Define PropTypes and modified file structure in product folder
description:
1. Define PropType for Product used in CatalogGrid & CatalogGridItem
2. Use file name as main file for code logic, and index just exporting
