+++ src/components/product/Price/Price.test.js
